Feat : 임시

+++ app/src/main/java/com/sas/companymanagement/ui/artist/db/ArtistRepository.kt
@@ -37,7 +37,7 @@ class ArtistRepository(application: Application) {
         }
     }
 
-    fun deleteArtist(id: Int) {
+    fun deleteArtist(id: Long) {
         coroutineScope.launch(Dispatchers.IO) {
             artistDao.deleteArtist(id)
         }
@@ -51,7 +51,7 @@ class ArtistRepository(application: Application) {
         }
     }
 
-    fun findArtist(id: Int) {
+    fun findArtist(id: Long) {
         coroutineScope.launch(Dispatchers.Main) {
             searchResults.value = coroutineScope.async(Dispatchers.IO) {
                 return@async artistDao.findArtist(id)
